+++ CHANGELOG.md
@@ -6,7 +6,9 @@
 
 - Increase "thread lock timer interval" for outgoing messages from 0.8 *period to 2.5* period. This should prevent the issue where messages are picked up by multiple threads if a process took too long to finish. (E.g. mail server down, and message sending is retried multiple times.)
 - If message sending failed, reset `lastSendAttempt` to 0, so that the message can be sent again in the next process.
+- If sending outgoing messages takes too long, skip sending remaining messages for the batch and try again in the next process.
 - Improved error messages for handling outgoing messages.
+- When updating an auto message schedule, use "upsert" with false -> this prevents an unexcepted behaviour where CosmosDB would return null when using FindOneAndReplace method.
 
 ## [v1.5.0] - 2024-01-08